requirements.txt
python-dotenv
openai
langchain-openai
langgraph
tiktoken
google-generativeai
langchain
PyPDF2
faiss-cpu
streamlit
-e .